--- src/main/java/com/example/springbootassignment/domain/Role.java
@@ -1,5 +1,0 @@
-package com.example.springbootassignment.domain;
-
-public enum Role {
-    USER, ADMIN
-}
