+++ src/main/java/com/example/springbootassignment/domain/comment/entity/Comment.java
@@ -1,18 +1,18 @@
-package com.example.springbootassignment.domain;
+package com.example.springbootassignment.domain.comment.entity;
 
+import com.example.springbootassignment.domain.post.entity.Post;
+import com.example.springbootassignment.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
-@Table(name = "post")
+@Table(name = "comment")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Post {
+public class Comment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -22,20 +22,12 @@ public class Post {
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
-    @Column(nullable = false, length = 100)
-    private String title;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id", nullable = false)
+    private Post post;
 
     @Column(nullable = false, columnDefinition = "TEXT")
     private String content;
-
-    @Column(length = 1000)
-    private String thumbnail;
-
-    @Column(name = "view_count", nullable = false)
-    private Integer viewCount = 0;
-
-    @Column(name = "comment_count", nullable = false)
-    private Integer commentCount = 0;
 
     @CreationTimestamp
     @Column(name = "created_at", updatable = false)
@@ -47,7 +39,4 @@ public class Post {
 
     @Column(name = "deleted_at")
     private LocalDateTime deletedAt;
-
-    @OneToMany(mappedBy = "post", cascade = CascadeType.ALL)
-    private List<Comment> comments = new ArrayList<>();
 }
