+++ src/main/java/com/example/springbootassignment/domain/post/entity/Post.java
@@ -1,16 +1,20 @@
-package com.example.springbootassignment.domain;
+package com.example.springbootassignment.domain.post.entity;
 
+import com.example.springbootassignment.domain.user.entity.User;
+import com.example.springbootassignment.domain.comment.entity.Comment;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
-@Table(name = "comment")
+@Table(name = "post")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment {
+public class Post {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -20,12 +24,20 @@ public class Comment {
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
-    private Post post;
+    @Column(nullable = false, length = 100)
+    private String title;
 
     @Column(nullable = false, columnDefinition = "TEXT")
     private String content;
+
+    @Column(length = 1000)
+    private String thumbnail;
+
+    @Column(name = "view_count", nullable = false)
+    private Integer viewCount = 0;
+
+    @Column(name = "comment_count", nullable = false)
+    private Integer commentCount = 0;
 
     @CreationTimestamp
     @Column(name = "created_at", updatable = false)
@@ -37,4 +49,7 @@ public class Comment {
 
     @Column(name = "deleted_at")
     private LocalDateTime deletedAt;
+
+    @OneToMany(mappedBy = "post", cascade = CascadeType.ALL)
+    private List<Comment> comments = new ArrayList<>();
 }
