+++ src/main/java/com/example/springbootassignment/domain/user/entity/Role.java
@@ -1,0 +1,5 @@
+package com.example.springbootassignment.domain.user.entity;
+
+public enum Role {
+    USER, ADMIN
+}
